tests: Consolidate resolveMultiLocation tests
parseLocationStrToLocation was previously consolidated with
xcmCreator.resolveMultiLocation
